+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Bug fixes and improvements
 
+
+- server: add warning log for missing admin secret
+
 ## v2.10.0-beta.1
 
 ### Introducing Apollo Federation v1 support (experimental)